update the mod name
