+++ proxy_agent_shared/src/etw/application.rs
@@ -1,4 +1,12 @@
+// Copyright (c) Microsoft Corporation
+// SPDX-License-Identifier: MIT
+//! This module provides functionality for Windows Application event
+//! // logging using the Windows Event Log API.
+//! // It allows registering an event source, writing logs, and reading events
+//! // from the Application Event Log.
+
 use crate::error::Error;
+use crate::etw::Event;
 use crate::logger::LoggerLevel;
 use crate::result::Result;
 use chrono::DateTime;
@@ -8,19 +16,25 @@ use std::iter::once;
 use std::os::windows::ffi::OsStrExt;
 use std::os::windows::ffi::OsStringExt;
 use windows_sys::core::PWSTR;
+use windows_sys::Win32::Foundation::GetLastError;
 use windows_sys::Win32::Foundation::{ERROR_NO_MORE_ITEMS, HANDLE};
 use windows_sys::Win32::System::EventLog::{
-    DeregisterEventSource, EvtClose, EvtNext, EvtQuery, EvtQueryReverseDirection, EvtRender,
-    EvtRenderEventXml, RegisterEventSourceW, ReportEventW, EVT_HANDLE,
-};
+    DeregisterEventSource, RegisterEventSourceW, ReportEventW,
+}; // advapi32.dll
+use windows_sys::Win32::System::EventLog::{EvtClose, EvtNext, EvtQuery, EvtRender, EVT_HANDLE}; // wevtapi.dll
 use windows_sys::Win32::System::EventLog::{
-    EVENTLOG_ERROR_TYPE, EVENTLOG_INFORMATION_TYPE, EVENTLOG_WARNING_TYPE, REPORT_EVENT_TYPE,
+    EvtQueryReverseDirection, EvtRenderEventXml, EVENTLOG_ERROR_TYPE, EVENTLOG_INFORMATION_TYPE,
+    EVENTLOG_WARNING_TYPE, REPORT_EVENT_TYPE,
 };
 
+/// Converts a string to a wide character vector (u16).
+/// This is used to convert Rust strings to the format required by Windows API functions.
 fn to_wide(s: &str) -> Vec<u16> {
     OsStr::new(s).encode_wide().chain(once(0)).collect()
 }
 
+/// Converts a `LoggerLevel` to a `REPORT_EVENT_TYPE`.
+/// This function maps the logging levels to the corresponding Windows Event Log types.
 fn to_event_level(level: LoggerLevel) -> REPORT_EVENT_TYPE {
     match level {
         LoggerLevel::Trace => EVENTLOG_INFORMATION_TYPE,
@@ -92,7 +106,6 @@ impl Drop for ApplicationEventWritter {
     }
 }
 
-use windows_sys::Win32::Foundation::GetLastError;
 pub struct WindowsEventReader {
     query_handle: EVT_HANDLE,
     current_event: EVT_HANDLE,
@@ -276,73 +289,6 @@ impl Iterator for WindowsEventReader {
             )))),
         }
     }
-}
-
-use serde_derive::{Deserialize, Serialize};
-
-#[derive(Debug, Deserialize, Serialize)]
-#[serde(rename = "Event")]
-pub struct Event {
-    #[serde(rename = "System")]
-    pub system: System,
-    #[serde(rename = "EventData")]
-    pub event_data: EventData,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct System {
-    #[serde(rename = "Provider")]
-    pub provider: Provider,
-    #[serde(rename = "EventID")]
-    pub event_id: u32,
-    #[serde(rename = "Version")]
-    pub version: u8,
-    #[serde(rename = "Level")]
-    pub level: u8,
-    #[serde(rename = "Task")]
-    pub task: u8,
-    #[serde(rename = "Opcode")]
-    pub opcode: u8,
-    #[serde(rename = "Keywords")]
-    pub keywords: String,
-    #[serde(rename = "TimeCreated")]
-    pub time_created: TimeCreated,
-    #[serde(rename = "EventRecordID")]
-    pub event_record_id: u64,
-    #[serde(rename = "Execution")]
-    pub execution: Execution,
-    #[serde(rename = "Channel")]
-    pub channel: String,
-    #[serde(rename = "Computer")]
-    pub computer: String,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct Provider {
-    #[serde(rename = "@Name", skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(rename = "@EventSourceName", skip_serializing_if = "Option::is_none")]
-    pub event_source_name: Option<String>,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct TimeCreated {
-    #[serde(rename = "@SystemTime", skip_serializing_if = "Option::is_none")]
-    pub system_time: Option<String>,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct Execution {
-    #[serde(rename = "@ProcessID")]
-    pub process_id: u32,
-    #[serde(rename = "@ThreadID")]
-    pub thread_id: u32,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct EventData {
-    #[serde(rename = "Data")]
-    pub data: Vec<String>,
 }
 
 #[cfg(test)]
